lan dau comit file lan sua sau

+++ src/bien/diemTB.java
@@ -5,14 +5,14 @@ public class diemTB {
     public static void main(String[] args) {
         Scanner scanner = new Scanner(System.in);
         System.out.println("nhap vao diem toan:");
-        int diemToan = scanner.nextInt();
+        float diemToan = scanner.nextInt();
         System.out.println("nhap vao diem ly:");
-        int diemLy = scanner.nextInt();
+        float diemLy = scanner.nextInt();
         System.out.println("nhap vao diem hoa:");
-        int diemHoa = scanner.nextInt();
-        int diemTB =(diemHoa + diemLy + diemToan)/3;
+        float diemHoa = scanner.nextInt();
+        float diemTB =(diemHoa + diemLy + diemToan)/3;
         System.out.println("nhap vao diem trung binh:" + diemTB);
-        int tong =(diemHoa + diemLy + diemToan);
+        float tong =(diemHoa + diemLy + diemToan);
         System.out.println("nhap vao diem trung binh:" + tong);
     }
 }
